@@ -17,9 +17,10 @@
     "antd": "^5.2.3",
     "firebase": "^9.17.1",
     "firebase-admin": "^11.5.0",
-    "next": "13.1.6",
+    "next": "^13.2.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-router-dom": "^6.8.2",
     "typescript": "4.9.5"
   }
 }
